[reorg] Move AlternateSubst to new layout.
