+++ src/meson.build
@@ -99,6 +99,9 @@ hb_base_sources = files(
   'OT/Layout/GSUB/SingleSubst.hh',
   'OT/Layout/GSUB/MultipleSubstFormat1.hh',
   'OT/Layout/GSUB/MultipleSubst.hh',
+  'OT/Layout/GSUB/AlternateSubstFormat1.hh',
+  'OT/Layout/GSUB/AlternateSubst.hh',
+  'OT/Layout/GSUB/AlternateSet.hh',
   'hb-ot-layout-gsubgpos.hh',
   'hb-ot-layout-jstf-table.hh',
   'hb-ot-layout.cc',
